pyproject.toml
[tool.poetry]
name = "mvp_coding_challenge"
version = "0.1.0"
description = ""
authors = ["George Pamfilis <gpamfilis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
Flask = "^2.2.2"
SQLAlchemy = "^1.4.43"
flask-sqlalchemy = "^3.0.2"
marshmallow = "^3.18.0"
Flask-Migrate = "^3.1.0"
python-dotenv = "^0.21.0"
Flask-WTF = "^1.0.1"
Flask-Cors = "^3.0.10"
psycopg2-binary = "^2.9.5"
itsdangerous = "2.0.1"
pytest = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
